+++ treesapp/__init__.py
@@ -1,4 +1,4 @@
 #! /usr/bin/env python
 """A Python package for gene-centric taxonomic and functional classification using phylogenetic placement."""
 name = "treesapp"
-__version__ = "0.10.1"
+__version__ = "0.10.2"