+++ treesapp/classy.py
@@ -695,6 +695,7 @@ class Updater(TreeSAPP):
         self.cluster_input = ""  # Used only if resolve is True
         self.clusters_prefix = ""  # Used only if resolve is True
         self.updated_refpkg_path = ""
+        self.training_dir = ""
         # self.rank_depth_map = None
         self.prop_sim = 1.0
         self.min_length = 0  # The minimum sequence length for a classified sequence to be included in the refpkg
@@ -702,7 +703,8 @@ class Updater(TreeSAPP):
 
         # Stage names only holds the required stages; auxiliary stages (e.g. RPKM, update) are added elsewhere
         self.stages = {0: ModuleFunction("lineages", 0),
-                       1: ModuleFunction("rebuild", 1)}
+                       1: ModuleFunction("rebuild", 1),
+                       2: ModuleFunction("train", 2)}
 
     def decide_stage(self, args):
         """
@@ -711,10 +713,13 @@ class Updater(TreeSAPP):
         This function ensures all the required inputs are present for beginning at the desired first stage,
         otherwise, the pipeline begins at the first possible stage to continue and ends once the desired stage is done.
 
-        If a
         :return: None
         """
         self.acc_to_lin = self.var_output_dir + os.sep + "accession_id_lineage_map.tsv"
+
+        if not self.input_sequences:
+            self.change_stage_status("train", False)
+
         self.validate_continue(args)
         return
 
@@ -772,15 +777,16 @@ class Creator(TreeSAPP):
         self.training_dir = ""
 
         # Stage names only holds the required stages; auxiliary stages (e.g. RPKM, update) are added elsewhere
-        self.stages = {0: ModuleFunction("search", 0),
-                       1: ModuleFunction("lineages", 1),
-                       2: ModuleFunction("clean", 2),
-                       3: ModuleFunction("cluster", 3),
-                       4: ModuleFunction("build", 4),
-                       5: ModuleFunction("evaluate", 5),
-                       6: ModuleFunction("support", 6),
-                       7: ModuleFunction("train", 7),
-                       8: ModuleFunction("update", 8)}
+        self.stages = {0: ModuleFunction("deduplicate", 0),
+                       1: ModuleFunction("search", 1),
+                       2: ModuleFunction("lineages", 2),
+                       3: ModuleFunction("clean", 3),
+                       4: ModuleFunction("cluster", 4),
+                       5: ModuleFunction("build", 5),
+                       6: ModuleFunction("evaluate", 6),
+                       7: ModuleFunction("support", 7),
+                       8: ModuleFunction("train", 8),
+                       9: ModuleFunction("update", 9)}
 
     def decide_stage(self, args):
         """
@@ -789,9 +795,11 @@ class Creator(TreeSAPP):
         This function ensures all the required inputs are present for beginning at the desired first stage,
         otherwise, the pipeline begins at the first possible stage to continue and ends once the desired stage is done.
 
-        If a
         :return: None
         """
+        if not args.dedup:
+            self.change_stage_status("deduplicate", False)
+
         if not args.profile:
             self.change_stage_status("search", False)
         if args.pc: